Refactor XML serialization logic into SmithyXmlPayloadCodec
- Moved serialization and deserialization logic from SmithyXmlSerializer to SmithyXmlPayloadCodec.
- Implemented methods for creating XML elements, writing values, and handling various Smithy shapes.
- Removed the SmithyXmlSerializer class as its functionality has been integrated into SmithyXmlPayloadCodec.
- Enhanced error handling and added support for XML attributes and flattened structures.
